Move `_` to `"_"` transformation to lower layer
